+++ CHANGELOG.md
@@ -12,12 +12,14 @@ __! Be certain to carefully back up any critical user data before installing thi
     - Can be dropped on another item to add the associated property.
     - To change the linked property, drop any property on the item pattern sheet.
 - Added an optional "Specificity" technique type to serve as a catch-all (by request).
-- Added Mantis Clan compendium entries
+- Added Mantis Clan compendium entries.
 - Fix : rnkMessage not passing on actor object for NPCs (thanks to Bragma).
-- Fix : The "Crescent Moon Style" technique rank from 4 to 2
+- Fix : The "Crescent Moon Style" technique rank from 4 to 2.
+- Fix : Drop an advancement on a PC/NPC sheet now correctly add the bonus to the Actor (ex Air +1).
 - QoL : RnK button is now black in chat if no actions are left in roll.
 - QoL : Added symbols legend in RnK dialog as reminder.
 - QoL : Added "(x Max)" display in RnK picker for max number of dice to keep (thanks to Bragma).
+- Others minor optimizations (ex: 20q saving multiple item at once).
 
 ## 1.2.1 - Praised be Firefox
 - Fix dice swap on firefox that overflowed on the top and bottom of the RnK dialog